refactor: :recycle: refactor teams department page
